Implement a track segment abstraction
Add a TrackSegment game object that houses one or multiple road segments
and handles rendering areas between the given road segments, as well as
areas left and right to all the road segments in the track segment.
These areas can either be single color or textured and the polygon count
of these areas can be restricted.